requirements.txt
munkres>=1.1.2
filterpy>=1.4.4
opencv-contrib-python>=4.1.2.30
billiard>=3.6.1.0
importlib_resources>=1.0.2
tqdm>=4.40.2
mysql-connector-python
